http://trac.atomiclabs.com/ticket/481
After a 304 response was sent (with no content length), HTTPParser was failing
to parse any additional requests contained within the TCP transaction.  This
bug was introduced when we add the logic in SnifferReactor to wait for ack
packets for timing metrics.

Problem was caused by checking HTTPMessage::isContentLengthImplied() before
setting the status code for HTTPResponse objects, which is required for the
call to work properly.

This fix moves the logic for setting status code (and other fields derived
solely from header parsing) from the HTTPMessage::finish() function into
HTTPMessage::finishHeaderParsing().